fix: resolve Kotlin migration conflicts and compile errors
