+++ src/main/kotlin/com/likelionknu/applyserver/application/data/repository/ApplicationRepository.kt
@@ -1,7 +1,6 @@
 package com.likelionknu.applyserver.application.data.repository
 
 import com.likelionknu.applyserver.admin.data.dto.response.AdminRecruitSummaryResponseDto
-import com.likelionknu.applyserver.application.data.entity.Application
 import com.likelionknu.applyserver.auth.data.entity.User
 import com.likelionknu.applyserver.auth.data.enums.ApplicationStatus
 import com.likelionknu.applyserver.recruit.data.entity.Recruit
