+++ src/main/kotlin/com/likelionknu/applyserver/application/service/ApplicationFinalSubmitService.kt
@@ -1,7 +1,6 @@
 package com.likelionknu.applyserver.application.service
 
 import com.likelionknu.applyserver.application.data.dto.request.FinalSubmitRequestDto
-import com.likelionknu.applyserver.application.data.entity.Application
 import com.likelionknu.applyserver.application.data.entity.RecruitAnswer
 import com.likelionknu.applyserver.application.data.exception.EmptyAnswerException
 import com.likelionknu.applyserver.application.data.exception.InvalidApplicationQuestionException
@@ -13,7 +12,7 @@ import com.likelionknu.applyserver.application.data.exception.UserNotFoundExcept
 import com.likelionknu.applyserver.application.data.repository.ApplicationRepository
 import com.likelionknu.applyserver.application.data.repository.RecruitAnswerRepository
 import com.likelionknu.applyserver.auth.data.enums.ApplicationStatus
-import com.likelionknu.applyserver.auth.repository.UserRepository
+import com.likelionknu.applyserver.auth.data.repository.UserRepository
 import com.likelionknu.applyserver.discord.service.DiscordNotificationService
 import com.likelionknu.applyserver.recruit.data.repository.RecruitContentRepository
 import com.likelionknu.applyserver.recruit.data.repository.RecruitRepository
