+++ src/main/kotlin/com/likelionknu/applyserver/application/service/ApplicationMailService.kt
@@ -1,6 +1,5 @@
 package com.likelionknu.applyserver.application.service
 
-import com.likelionknu.applyserver.application.data.entity.Application
 import com.likelionknu.applyserver.application.data.repository.ApplicationRepository
 import com.likelionknu.applyserver.auth.data.entity.User
 import com.likelionknu.applyserver.auth.data.enums.ApplicationStatus
