+++ src/main/kotlin/com/likelionknu/applyserver/application/service/ApplicationService.kt
@@ -3,13 +3,12 @@ package com.likelionknu.applyserver.application.service
 import com.likelionknu.applyserver.application.data.dto.request.ApplicationDraftSaveRequest
 import com.likelionknu.applyserver.application.data.dto.response.ApplicationAnswerResponseDto
 import com.likelionknu.applyserver.application.data.dto.response.ApplicationInfoResponseDto
-import com.likelionknu.applyserver.application.data.entity.Application
 import com.likelionknu.applyserver.application.data.exception.ApplicationNotFoundException
 import com.likelionknu.applyserver.application.data.exception.ApplicationStateException
 import com.likelionknu.applyserver.application.data.exception.UserNotFoundException
 import com.likelionknu.applyserver.application.data.repository.ApplicationRepository
 import com.likelionknu.applyserver.auth.data.enums.ApplicationStatus
-import com.likelionknu.applyserver.auth.repository.UserRepository
+import com.likelionknu.applyserver.auth.data.repository.UserRepository
 import com.likelionknu.applyserver.common.response.ErrorCode
 import com.likelionknu.applyserver.common.response.GlobalException
 import com.likelionknu.applyserver.discord.service.DiscordNotificationService
