+++ src/main/kotlin/com/likelionknu/applyserver/discord/service/DiscordFeignClient.kt
@@ -1,4 +1,4 @@
-package com.likelionknu.applyserver.discord.data.service
+package com.likelionknu.applyserver.discord.service
 
 import com.likelionknu.applyserver.discord.data.dto.DiscordMessageDto
 import org.springframework.cloud.openfeign.FeignClient
